Add postprocessing parameters field to Page class (implements #140)
For now, we only store 'crop' and 'rotate' values. The former should be
a 4-tuple of (x, y, width, height) of the area to be cropped, the latter
an integer that is a multiple of 90.
The responsibility for cropping has accordingly been moved from the
Worfklow class (were it never should have been in the first place...) to
the new 'croprotate' plugin, which supersedes (and is based on) the
'autorotate' plugin, so _UPDATE YOUR CONFIGURATION_. It uses the
processing parameters from each page and performs the operations
accordingly.
Also removed was the setting of the EXIF rotation in the device classes.
This has become unneccessary, since we can directly set the 'rotate' key
in the page's postprocessing parameters. This is more robust than the
admittedly clunky hack with the EXIF orientation tag and should also
help with running spreads on other platforms, since we no longer have to
write EXIF tags.